Properly check if {SYS_SELECT, TIME}_WITH_SYS_TIME should be set
The usage of "check_include_files" introduced in commit 9baa8f
is not appropriate in that particular case. The variable was set to True
if both header were present on the system.

Instead, a compile test checking that both headers can be included
together in a same file is done.