Create versioned symlinks on install to match autoconf python install.
Use both DESTDIR and CMAKE_INSTALL_PREFIX to specify the working directory.
That way install location can be appropriately defined at both configure
and build time. See [1] and [2] for more details.

[1] http://cmake.3232098.n2.nabble.com/Confusion-about-CMAKE-INSTALL-PREFIX-td7577447.html#a7577461
[2] http://www.cmake.org/Wiki/CMake_FAQ#Does_CMake.27s_.22make_install.22_support_DESTDIR.3F

The symlinks are created as:

  python${PY_VERSION_MAJOR}.${PY_VERSION_MINOR} -> python
  python${PY_VERSION_MAJOR} -> python

which is DIFFERENT from the symlink chain created in the corresponding
autoconf file 'Makefile.pre.in':

// ------------------
[...]
bininstall:	altbininstall
	[...]
	(cd $(DESTDIR)$(BINDIR); $(LN) -s python2$(EXE) $(PYTHON))
	-rm -f $(DESTDIR)$(BINDIR)/python2$(EXE)
	(cd $(DESTDIR)$(BINDIR); $(LN) -s python$(VERSION)$(EXE) python2$(EXE))
	-rm -f $(DESTDIR)$(BINDIR)/python2-config
	[...]
// ------------------

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>

Fixes #26