Fix get_target_property warning related to non existent target
With CMake >= 3.0.0, the developer is warned when his code attempt to get
a property of an non-existent target. This is a good thing. In this
particular case, the code check if the returned value is set or not to
check if the "library" is an imported target or path to a library. For that
reason, it is safe to ignore the warning.

The particular problem has been reported upstream:
  http://www.cmake.org/Bug/view.php?id=15008

Waiting a fix is integrated upstream, this commit apply the fixes to
the version used for CMake version <= 2.8.7 and make sure to is used for
CMake version having the policy CMP0045.

Since the version currently in CMake and the one available in this
repository are not different (beside of some minor style fixes), there
are no issue using the local version of CMake >- 3.0.0