Simplify if() conditions on check result variables to fix CMP0054 warnings
Remove use of an old hack that takes advantage of the auto-dereference
behavior of the if() command to detect if a variable is defined.  The
hack has the form:

 if("${VAR} MATCHES "^${VAR}$")

where "${VAR}" is a macro argument reference.  Use if(NOT DEFINED) instead.
This also avoids warnings for CMake Policy CMP0054.


This commit fixes the following warnings:

CMake Warning (dev) at cmake/PlatformTest.cmake:69 (if):
  Policy CMP0054 is not set: Only interpret if() arguments as variables or
  keywords when unquoted.  Run "cmake --help-policy CMP0054" for policy
  details.  Use the cmake_policy command to set the policy and suppress this
  warning.

  Quoted variables like "DOUBLE_IS_LITTLE_ENDIAN_IEEE754" will no longer be
  dereferenced when the policy is set to NEW.  Since the policy is not set
  the OLD behavior will be used.
Call Stack (most recent call first):
  cmake/ConfigureChecks.cmake:642 (python_platform_test_run)
  CMakeLists.txt:104 (include)
This warning is for project developers.  Use -Wno-dev to suppress it.

CMake Warning (dev) at cmake/PlatformTest.cmake:10 (IF):
  Policy CMP0054 is not set: Only interpret if() arguments as variables or
  keywords when unquoted.  Run "cmake --help-policy CMP0054" for policy
  details.  Use the cmake_policy command to set the policy and suppress this
  warning.

  Quoted variables like "HAVE_GCC_ASM_FOR_X87_COMPILED" will no longer be
  dereferenced when the policy is set to NEW.  Since the policy is not set
  the OLD behavior will be used.
Call Stack (most recent call first):
  cmake/ConfigureChecks.cmake:693 (python_platform_test)
  CMakeLists.txt:104 (include)
This warning is for project developers.  Use -Wno-dev to suppress it.