Add option INSTALL_WINDOWS_TRADITIONAL
By default OFF, this option allows to install python executable, libraries
and headers in a directory hierarchy matching the layout of the official
distribution.

INSTALL_WINDOWS_TRADITIONAL:OFF

 Executable ..: <BUILD_DIR>/bin               <INSTALL_DIR>/bin
 Libraries ...: <BUILD_DIR>/Lib/lib-dynload   <INSTALL_DIR>/Lib/lib-dynload

INSTALL_WINDOWS_TRADITIONAL:ON

 Executable ..: <BUILD_DIR>/bin               <INSTALL_DIR>/.
 Libraries ...: <BUILD_DIR>/Lib/lib-dynload   <INSTALL_DIR>/DLLs

Note: Moving forward, we may revisit the use of this option and
instead ensure the layout is always consistent with the official
distribution.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>