Allow configuration providing the patch version. Fixes #57
This commit allows selection of the python version passing the option
PY_VERSION_PATCH.

For example, to download, configure and build version 2.7.9, the following
could be done:

  mkdir python-cmake-buildsystem-bld
  cd  python-cmake-buildsystem-bld
  cmake -DPY_VERSION_PATCH=9 ../python-cmake-buildsystem

Additionally, note that it is still possible to configure python by

(1) using the default value. In that case, Python 2.7.8 will be
configured and built. This is an out-of-source build.

For example:

  mkdir python-cmake-buildsystem-bld
  cd  python-cmake-buildsystem-bld
  cmake ../python-cmake-buildsystem

(2) providing the path to an arbitrary source directory for any
version from 2.7.3 to 2.7.10. In that case, the build system will
auto-discover the python version reading it from ${SRC_DIR}/Include/patchlevel.h.
This is an out-of-source build..

For example:

  mkdir python-cmake-buildsystem-bld
  cd  python-cmake-buildsystem-bld
  cmake -DSRC_DIR:PATH=/path/to/Python-Source ../python-cmake-buildsystem

(3) configuring from the Python source tree directly. In that case,
similarly to what was described in (2), the python version will also be
auto-discovered.

For example:

  tar -xzvf Python-2.7.9.tar.gz
  cd Python-2.7.9
  cmake ../python-cmake-buildsystem

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>