Fix "pyexpat" import error when "_sqlite3" and "pyexpat" where "builtin"
Since "_sqlite3" extension was specifying "MODULE_NAME" as a DEFINITIONS,
it ended up in the global list of definitions used unconditionally
to compile statically the two extensions into the libpython library.

The side effect was that "pyexpat.h" was associating the module pyexpat
with the name "sqlite3" because of the MODULE_NAME definition test in
"pyexpat.h".

By explicitly setting the DEFINITIONS on the corresponding files, this
commit avoid MODULE_NAME to be associated with all extensions source files
statically linked into libpython.

The error was:

$ python -c "import pyexpat"
SystemError: _PyImport_FixupExtension: module pyexpat not loaded

This commit reverts a39a713 and 3a08d98.