Add additional locations to check for python source
In addition to along side the binary dir, some additional folders in the
source dir are checked.  This essentially allows for the CMake source
and upstream python source to be tar'd up together, but still keep them
distinctly separate.  Previously this could only be done extracting the
CMake source into the top-level folder of the upstream python source.