Fix `exec_prefix` error occurring when no extension are built shared
This commit adds an empty `PYTHONHOME/lib-dynload` directory to
the build and install tree to respectively fix the following errors:

Build error:

  [...]
  Relocate _sysconfigdata.py and update pybuilddir.txt
  Could not find platform dependent libraries <exec_prefix>
  Consider setting $PYTHONHOME to <prefix>[:<exec_prefix>]

Runtime error:

  cd /home/jchris/python-cmake-buildsystem-install
  $ ./bin/python -c "print('Hello')"
  Could not find platform dependent libraries <exec_prefix>
  Consider setting $PYTHONHOME to <prefix>[:<exec_prefix>]
  Hello

To reproduce the error, the build system was configured using
the following options:

  cmake -DCMAKE_INSTALL_PREFIX=/home/jchris/python-cmake-buildsystem-install \
    -DBUILD_EXTENSIONS_AS_BUILTIN:BOOL=1 \
    -DENABLE_CTYPES_TEST=0 -DENABLE_TESTCAPI=0 \
    ../python-cmake-buildsystem

It turns out that if all modules are `BUILTIN`, there will be no
`lib-dynload` folder. `bin/python` uses this to auto-determine
the 'exec_prefix' and properly generate the `_sysconfigdata.py`
so we create an empty directory just in case to make sure it
will always exist.

Reviewed-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>