Fix OS X build when CMAKE_BUILD_WITH_INSTALL_RPATH is enabled
When building Python 2.7.4+, python is run from the build tree as part of the
build process.  When BUILD_SHARED and CMAKE_BUILD_WITH_INSTALL_RPATH are enabled
the python executable is unable to find the shared library.  The build fails
with an error like:

    [100%] Relocate _sysconfigdata.py and update pybuilddir.txt
    dyld: Library not loaded: /path/to/python-install/lib/libpython2.7.dylib
      Referenced from: /path/to/python-build/bin/python
      Reason: image not found
    /bin/sh: line 1: 12570 Trace/BPT trap          /path/to/python-build/bin/python -E -S -m sysconfig --generate-posix-vars
    make[2]: *** [bin/pybuilddir.txt] Error 133
    make[1]: *** [CMakeBuild/python/CMakeFiles/update_sysconfig.dir/all] Error 2
    make: *** [all] Error 2

The same error shows when running the tests.

This commit sets the DYLD_LIBRARY_PATH so that the executable finds the shared
library when building and when running the tests.