Check build fails on MacOSX when build as shared with build_with_install_rpath
THIS COMMIT SHOUD NOT BE INTEGRATED INTO MASTER

The role of the commit is to confirm the build fail and will be effectively
fixed by PR #97.

It also test the feature introduced in 0f8541e9c42 teaching the driver
script how to extract additional CMake configuration arguments from the
commit message. This new feature provides a simple way to test some
options without having to explicitly change the CI build matrix.

This is special awesome comment:

[cmake -DCMAKE_BUILD_WITH_INSTALL_RPATH=ON]