Add WITH_STATIC_DEPENDENCIES option (available only on Unix)
This will restrict extensions and external dependencies to static
libraries.

Enabling the option is roughly equivalent to configuring with
  * CMAKE_FIND_LIBRARY_SUFFIXES set to ".a"
  * BUILD_EXTENSIONS_AS_BUILTIN set to ON

To support the option, this commit applies the following changes:

(1) instead of explicitly looking for the dl library, it uses
CMAKE_DL_LIBS to initialize HAVE_LIBDL. Considering that CMAKE_DL_LIBS
is set to the library containing the dlopen and dlclose symbols. This will
allow to build statically while linking against the "dl" shared library and
avoid the warning:

  warning: Using 'dlopen' in statically linked applications requires at runtime the shared libraries from the glibc version used for linking

(2) look up library "gpm" and "tinfo" required to statically link against
ncurses (at least on Ubuntu 15.10). It avoids link error about missing
symbols like:
  * wtimeout, resetty, ... provided by "tinfo" library,
  * enable_gpm_mouse, mouse_activate, ... provided by "gpm" library.

Co-authored-by: Chuck Atkins <chuck.atkins@kitware.com>