pyconfig: Check for _LARGEFILE_SOURCE definition inconsistency
When embedding Python in other application, it is not practical to
follow the official recommendations [1] asking to include "Python.h" first.

Failing to do so can also causes "redefinition" warnings (see previous
commit), to mitigate this while ensuring Python code is compiled
with the expected ABI [2], this commit checks if previous definitions
of _LARGEFILE_SOURCE are consistent with the value expected by CPython
build system.

[1] https://docs.python.org/3/c-api/intro.html#include-files
[2] http://bugs.python.org/issue1045893#msg22669