PythonApplyPatches: Use git executable it found, and check for older version of patch
Since on macOS, the version of patch executable available on the system
is too old to support rename and copy of files found in patches generated
using git, this commit update the PythonApplyPatches so that it also
consider git exectutable to apply patches and report a clear message
if only patch executable was found and it is too old.

This was verified on macOS 10.11.6 where patch 2.5.8 is available.