python3/sysmodule: Link against "version.lib"
Starting with Python3.5, the function GetFileVersionInfoW is used
in "sysmodule.c". See cpython/python@65e4cb1

Note that it started to be used in pylauncher in Python 3.3.

Finally, worth noting that following the link error that ultimately lead
to this fix, simply searching the Microsoft documentation wasn't enough.
Indeed, the documentation only references the name of the library (minicore.lib)
for Windows 8 and above. See [1].

[1] https://msdn.microsoft.com/en-us/library/ms647005%28v=VS.85%29.aspx?f=255&MSPPError=-2147217396

It is by looking at the module "PC/validate_ucrtbase.py" and some more
research that we realized "version.lib" was the correct library. See [2][3]

[2] http://clang-developers.42468.n3.nabble.com/mincore-lib-symbols-in-the-link-how-to-avoid-tp4057650.html
[3] https://github.com/poizan42/soswow64/commit/2f9e368c28476caff1ce2966f87e9122c422c804