cmake: Disable weak linking to "fix Symbol not found: _clock_gettime"
This commit was adapted from the work done by @sitsofe in
https://github.com/axboe/fio/pull/309/commits/ccf2d89d39b21bc8c7b497b40be5b82eadb80863

Description below was copied verbatim from the commit referenced above.

macOS 10.12 introduced support for clock_gettime() but when compiling
with Xcode 8 (or later) and targeting 10.11 (or older) the clock_gettime
symbol is now found at configure/compile time but referenced weakly.
Running the created binary on 10.11 results in a "dyld: lazy symbol
binding failed: Symbol not found: _clock_gettime" error.

Options for working around this issue include:

1. Turn references to functions that might be weak into compilation
   errors (-Werror=partial-availability).

2. Disable visibility of functions that might be weak at compile link
   time (-Wl,-no_weak_imports).

3. Change configure tests to directly check the targeted OSX version and
   fail the test if it is too old to have the required function.

4. Make an empty function declaration marked with
   __attribute__((weak_import)) that is used if the symbol would
   otherwise be unavailable. It is then possible to check if the symbol
   is NULL at runtime to determine availability.

This commit implements approach (2) when targeting the Darwin platform
after checking that the compiler/linker work with that option.

Co-authored-by: Sitsofe Wheeler <sitsofe@yahoo.com>