cmake: Fix undefined reference to gss_* symbols when libkrb5-dev package not installed
This commit is a follow-up of f929319a9 (cmake: Fix undefined reference to
gss_* symbols when WITH_STATIC_DEPENDENCIES is ON).

If the shared library "libgssapi_krb5.so" can't be found, the build-system
is updated to assume the static libraries are required.

Indeed, without installing the "libkrb5-dev" package, the shared library
"libgssapi_krb5.so" can't be found and the test checking if the tirpc
library was build with or without the gssapi feature can not be performed.

See #227 and #285