cmake: Remove duplicated sources between libpython and extensions
This commit removes the sources of the following extensions from
libpython/CMakeLists.txt and update extensions/CMakeLists.txt to
support either python 2 or 3.

It also update config_2.c to remove the hard-coded entries. These
are added automatically when the extensions are built-in,

Fix build error like the following by excluding timemodule.c from
the list of sources associated with _socket module. The extra file
was added to Python 2.7.4 in python/cpython@868f0aac3 (bpo-9090 : Take
the same approach for socketmodule as daytimemodule when it needs
support from timemodule) but corresponding changes is not found in
any of the Python 3.x versions.

  $ ninja extension_socket

  /path/to/Python-3.6.15/Modules/timemodule.c: In function ‘time_strftime’:
  /path/to/Python-3.6.15/Modules/timemodule.c:677:9: error: unknown type name ‘_Py_BEGIN_SUPPRESS_IPH’
           _Py_BEGIN_SUPPRESS_IPH
           ^
  /path/to/Python-3.6.15/Modules/timemodule.c:679:9: error: ‘_Py_END_SUPPRESS_IPH’ undeclared (first use in this function)
           _Py_END_SUPPRESS_IPH
           ^
  /path/to/Python-3.6.15/Modules/timemodule.c:679:9: note: each undeclared identifier is reported only once for each function it appears in
  /path/to/Python-3.6.15/Modules/timemodule.c:688:9: error: expected ‘;’ before ‘if’
           if (buflen > 0 || i >= 256 * fmtlen) {
           ^
  /path/to/Python-3.6.15/Modules/timemodule.c:678:9: warning: unused variable ‘buflen’ [-Wunused-variable]
           buflen = format_time(outbuf, i, fmt, &buf);
           ^
  /path/to/Python-3.6.15/Modules/timemodule.c:578:20: warning: unused variable ‘buflen’ [-Wunused-variable]
       size_t fmtlen, buflen;
                      ^
  /path/to/Python-3.6.15/Modules/timemodule.c:578:12: warning: variable ‘fmtlen’ set but not used [-Wunused-but-set-variable]
       size_t fmtlen, buflen;
              ^